refactor(types): replace hand-rolled parser with winnow combinators
Split format.rs (1500 lines) into format/mod.rs + format/selector.rs.
Replaced 178 lines of hand-rolled recursive descent parser with ~85
lines of winnow 0.7 combinators (separated, alt, delimited, repeat).
All 40 tests + 1 doc-test pass unchanged.
