+++ crates/rdlp-types/Cargo.toml
@@ -11,6 +11,7 @@ serde = { workspace = true }
 serde_json = { workspace = true }
 url = { workspace = true }
 regex = { workspace = true }
+winnow = { workspace = true }
 
 [dev-dependencies]
